Move findVerificationMethod to resolver.
