+++ src/main.js
@@ -1,5 +1,4 @@
 /*!
  * Copyright (c) 2020-2021 Digital Bazaar, Inc. All rights reserved.
  */
-export {DidDocument} from './DidDocument.js';
 export {DidResolver} from './DidResolver.js';
